move files and clean up unused files
