+++ src/components/ChatbotInterface.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { useLocation } from "react-router-dom";
-import { randomPhrase } from "../../utils/randomPhrase";
-import speak from "../../utils/textToSpeech";
-import getRandomPlantEmoji from "../../utils/getPlantEmoji";
-import { askAssistant } from "../../utils/openAiApi";
+import { randomPhrase } from "../utils/randomPhrase.js";
+import speak from "../utils/textToSpeech.js";
+import getRandomPlantEmoji from "../utils/getPlantEmoji.js";
+import { askAssistant } from "../utils/openAiApi.js";
 
 export default function Chatbot() {
   const [messages, setMessages] = useState([]);
